第六次提交

+++ app/src/main/java/com/example/viewpager/FooBean/FooHomeBean.java
@@ -266,7 +266,7 @@ public class FooHomeBean {
             private Integer id;
             private String name;
             private String list_pic_url;
-            private Integer retail_price;
+            private Double retail_price;
 
             public Integer getId() {
                 return id;
@@ -292,11 +292,11 @@ public class FooHomeBean {
                 this.list_pic_url = list_pic_url;
             }
 
-            public Integer getRetail_price() {
+            public Double getRetail_price() {
                 return retail_price;
             }
 
-            public void setRetail_price(Integer retail_price) {
+            public void setRetail_price(Double retail_price) {
                 this.retail_price = retail_price;
             }
         }
@@ -313,7 +313,7 @@ public class FooHomeBean {
             private Integer id;
             private String name;
             private String list_pic_url;
-            private Integer retail_price;
+            private Double retail_price;
             private String goods_brief;
 
             public Integer getId() {
@@ -340,11 +340,11 @@ public class FooHomeBean {
                 this.list_pic_url = list_pic_url;
             }
 
-            public Integer getRetail_price() {
+            public Double getRetail_price() {
                 return retail_price;
             }
 
-            public void setRetail_price(Integer retail_price) {
+            public void setRetail_price(Double retail_price) {
                 this.retail_price = retail_price;
             }
 
@@ -382,7 +382,7 @@ public class FooHomeBean {
             private String pic_url;
             private Integer sort_order;
             private Integer is_show;
-            private Integer floor_price;
+            private Double floor_price;
             private String app_list_pic_url;
             private Integer is_new;
             private String new_pic_url;
@@ -444,11 +444,11 @@ public class FooHomeBean {
                 this.is_show = is_show;
             }
 
-            public Integer getFloor_price() {
+            public Double getFloor_price() {
                 return floor_price;
             }
 
-            public void setFloor_price(Integer floor_price) {
+            public void setFloor_price(Double floor_price) {
                 this.floor_price = floor_price;
             }
 
@@ -516,7 +516,7 @@ public class FooHomeBean {
             private String item_pic_url;
             private String subtitle;
             private Integer topic_category_id;
-            private Integer price_info;
+            private Double price_info;
             private String read_count;
             private String scene_pic_url;
             private Integer topic_template_id;
@@ -580,11 +580,11 @@ public class FooHomeBean {
                 this.topic_category_id = topic_category_id;
             }
 
-            public Integer getPrice_info() {
+            public Double getPrice_info() {
                 return price_info;
             }
 
-            public void setPrice_info(Integer price_info) {
+            public void setPrice_info(Double price_info) {
                 this.price_info = price_info;
             }
 
@@ -683,7 +683,7 @@ public class FooHomeBean {
                 private Integer id;
                 private String name;
                 private String list_pic_url;
-                private Integer retail_price;
+                private Double retail_price;
 
                 public Integer getId() {
                     return id;
@@ -709,11 +709,11 @@ public class FooHomeBean {
                     this.list_pic_url = list_pic_url;
                 }
 
-                public Integer getRetail_price() {
+                public Double getRetail_price() {
                     return retail_price;
                 }
 
-                public void setRetail_price(Integer retail_price) {
+                public void setRetail_price(Double retail_price) {
                     this.retail_price = retail_price;
                 }
             }
